requirements.txt
asgiref==3.5.2
crispy-bootstrap5==0.6
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.1
django-active-link==0.1.8
django-ckeditor==6.5.1
django-countries==7.3.2
django-crispy-forms==1.14.0
django-discover-runner==1.0
django-environ==0.9.0
django-import-export==2.8.0
django-js-asset==2.0.0
django-model-utils==4.2.0
django-role-permissions==3.1.1
django-taggit==3.0.0
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
typing_extensions==4.3.0
tzdata==2022.2
xlrd==2.0.1
xlwt==1.3.0
